requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Speech-to-Text
openai-whisper>=20231117
faster-whisper>=0.10.0
soundfile>=0.12.1
numpy>=1.24.0

# Text-to-Speech
edge-tts>=6.1.9
openai>=1.3.0

# Audio Processing
pydub>=0.25.1
webrtcvad>=2.0.10

# SCL Integration
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.2
